% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddspls.R
\name{ddsPLS}
\alias{ddsPLS}
\title{ddsPLS}
\usage{
ddsPLS(
  X,
  Y,
  lambdas,
  n_B,
  minBootProp = 0,
  lowExplainedVariance = 0,
  NCORES = 1,
  errorMin = 1e-09,
  verbose = FALSE
)
}
\arguments{
\item{X}{matrix, the covariate matrix (n,p)}

\item{Y}{matrix, the response matrix (n,q)}

\item{lambdas}{vector, the to be tested values for lambda}

\item{n_B}{integer, the number of to be simulated bootstrap samples}

\item{minBootProp}{real, between 0 and 1, the minimum proportion of non null
components built to accept the current lambda value. Default to 0.0}

\item{lowExplainedVariance}{real, the minimum value of Q^2_B to accept the
current lambda value. Default to 0.0}

\item{NCORES}{integer, the number of cores used. Default to 1}

\item{errorMin}{real, not to be used}

\item{verbose}{boolean, whether to print current results}
}
\value{

}
\description{
The main function of the package. It allows both:
}
\details{
*  starting the ddsPLS algorithm, through the bootstrap work.
  *  Building the chosen model, of the class ddsPLS.
}
