% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.ddsPLS.R
\name{predict.ddsPLS}
\alias{predict.ddsPLS}
\title{Function to predict from ddsPLS objects}
\usage{
\method{predict}{ddsPLS}(x, X_test, ...)
}
\arguments{
\item{x}{A ddsPLS object}

\item{X_test}{Optimal number of components to be plotted}

\item{...}{Other parameters.}
}
\description{
Function to predict from ddsPLS objects
}
