% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.ddsPLS.R
\name{predict.ddsPLS}
\alias{predict.ddsPLS}
\title{Function to predict from ddsPLS objects}
\usage{
\method{predict}{ddsPLS}(
  x,
  X_test = NULL,
  doDiagnosis = T,
  legend.position = "topright",
  cex = 1,
  cex.text = 1,
  ...
)
}
\arguments{
\item{x}{ddsPLS object}

\item{X_test}{matrix, a test data-set. If is "NULL", the default value, the predicted values for the train test are returned}

\item{legend.position}{character. Where to put the legend.}

\item{cex}{float positive. Number indicating the amount by which plotting symbols should be scaled relative to the default. 1=default, 1.5 is 50% larger, 0.5 is 50% smaller, etc}

\item{cex.text}{float positive. Number indicating the amount by which plotting text elements should be scaled relative to the default. 1=default, 1.5 is 50% larger, 0.5 is 50% smaller, etc}

\item{...}{Other parameters}
}
\description{
Function to predict from ddsPLS objects
}
\seealso{
\code{\link{ddsPLS}}, \code{\link{plot.ddsPLS}}, \code{\link{summary.ddsPLS}}
}
