% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.ddsPLS.R
\name{predict.ddsPLS}
\alias{predict.ddsPLS}
\title{Function to predict from ddsPLS objects}
\usage{
\method{predict}{ddsPLS}(x, X_test = NULL, ...)
}
\arguments{
\item{x}{ddsPLS object}

\item{X_test}{matrix, a test data-set. If is "NULL", the default value, the predicted values for the train test are returned}

\item{...}{Other parameters}
}
\description{
Function to predict from ddsPLS objects
}
\seealso{
\code{\link{ddsPLS}}, \code{\link{plot.ddsPLS}}, \code{\link{summary.ddsPLS}}
}
